+++ src/main/java/roomescape/service/ReservationService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.dao.ReservationDao;
-import roomescape.dao.ReservationTimeDao;
-import roomescape.dao.ThemeDao;
+import roomescape.repository.ReservationRepository;
+import roomescape.repository.ReservationTimeRepository;
+import roomescape.repository.ThemeRepository;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationStatus;
 import roomescape.domain.ReservationTime;
@@ -20,25 +20,25 @@ import roomescape.exception.NotFoundException;
 
 @Service
 public class ReservationService {
-    private final ReservationDao reservationDao;
-    private final ReservationTimeDao reservationTimeDao;
-    private final ThemeDao themeDao;
+    private final ReservationRepository reservationRepository;
+    private final ReservationTimeRepository reservationTimeRepository;
+    private final ThemeRepository themeRepository;
 
-    public ReservationService(ReservationDao reservationDao, ReservationTimeDao reservationTimeDao, ThemeDao themeDao) {
-        this.reservationDao = reservationDao;
-        this.reservationTimeDao = reservationTimeDao;
-        this.themeDao = themeDao;
+    public ReservationService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository, ThemeRepository themeRepository) {
+        this.reservationRepository = reservationRepository;
+        this.reservationTimeRepository = reservationTimeRepository;
+        this.themeRepository = themeRepository;
     }
 
     public List<ReservationRankResponse> find(String name) {
-        return reservationDao.findByName(name)
+        return reservationRepository.findByName(name)
                 .stream()
                 .map(ReservationRankResponse::from)
                 .toList();
     }
 
     public List<ReservationResponse> findAll() {
-        return reservationDao.findAll()
+        return reservationRepository.findAll()
                 .stream()
                 .map(ReservationResponse::from)
                 .collect(Collectors.toList());
@@ -60,7 +60,7 @@ public class ReservationService {
         reservation.validateNotPast(request.date(), time);
         validateDuplicate(reservation);
 
-        Reservation saved = reservationDao.save(reservation);
+        Reservation saved = reservationRepository.save(reservation);
         return ReservationResponse.from(saved);
     }
 
@@ -84,10 +84,10 @@ public class ReservationService {
         newReservation.validateNotPast(request.date(), time);
         validateDuplicate(newReservation);
 
-        reservationDao.delete(id);
-        reservationDao.update(reservation.getDate(), reservation.getTheme().getId(), reservation.getTime().getId());
+        reservationRepository.delete(id);
+        reservationRepository.update(reservation.getDate(), reservation.getTheme().getId(), reservation.getTime().getId());
 
-        Reservation saved = reservationDao.save(newReservation);
+        Reservation saved = reservationRepository.save(newReservation);
 
         return ReservationResponse.from(saved);
     }
@@ -95,29 +95,29 @@ public class ReservationService {
     @Transactional
     public void delete(Long id) {
         Reservation reservation = getReservation(id);
-        reservationDao.delete(id);
+        reservationRepository.delete(id);
 
-        reservationDao.update(reservation.getDate(), reservation.getTheme().getId(), reservation.getTime().getId());
+        reservationRepository.update(reservation.getDate(), reservation.getTheme().getId(), reservation.getTime().getId());
 
     }
 
     private Reservation getReservation(long id) {
-        return reservationDao.findById(id)
+        return reservationRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("요청하신 예약을 찾을 수 없습니다."));
     }
 
     private ReservationTime getReservationTime(Long timeId) {
-        return reservationTimeDao.findTimeById(timeId)
+        return reservationTimeRepository.findTimeById(timeId)
                 .orElseThrow(() -> new NotFoundException("요청하신 시간 정보를 찾을 수 없습니다. 선택하신 시간이 정확한지 다시 한번 확인해 주세요."));
     }
 
     private Theme getTheme(Long themeId) {
-        return themeDao.findThemeById(themeId)
+        return themeRepository.findThemeById(themeId)
                 .orElseThrow(() -> new NotFoundException("요청하신 테마를 찾을 수 없습니다. 선택하신 테마가 정확한지 다시 한번 확인해 주세요."));
     }
 
     private void validateDuplicate(Reservation reservation) {
-        if (reservationDao.existsByDateAndThemeAndTimeAndName(
+        if (reservationRepository.existsByDateAndThemeAndTimeAndName(
                 reservation.getDate(),
                 reservation.getTheme().getId(),
                 reservation.getTime().getId(),
